fix(server/migration): make duration as int type instead of tinyint

+++ packages/server/src/databases/typeorm/migrations/1672946728513-CreateContentTable.ts
@@ -23,7 +23,7 @@ export class CreateContentTable1672946728513 implements MigrationInterface {
                     },
                     {
                         name: "duration",
-                        type: "tinyint",
+                        type: "int",
                         isNullable: false,
                     },
                     {
